ScopGraphPrinter: Swtiched from record to HTML-Like node labels
This has the greate advantage that every code line can have a different
backcolor. This is now used to show the profiling infromation inside of
the graph node labels. Hot lines have red backcolor.
Before only a complete basic block could had a backcolor derived from
the sum of the profiling percentage of every line of code the belonged
to that bascic block.
This will only work if the HTML node patch is applied to LLVM (see
https://github.com/tepelmann/llvm/commit/d1330321964b707e7c52655549b9e30c52d0f64e)